mobile_app/resources/requirements.txt
gitpython>=1.0.1
requests>=2.5.3
